refactor: 인수 테스트에서 공통적으로 사용되는 로직들 acceptanceTest로 이동
