+++ src/test/java/wooteco/subway/line/SectionAcceptanceTest.java
@@ -12,7 +12,6 @@ import wooteco.subway.AcceptanceTest;
 import wooteco.subway.auth.dto.TokenResponse;
 import wooteco.subway.line.dto.LineResponse;
 import wooteco.subway.line.dto.SectionDistanceRequest;
-import wooteco.subway.line.dto.SectionRequest;
 import wooteco.subway.station.dto.StationResponse;
 
 import java.util.Arrays;
@@ -21,7 +20,6 @@ import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static wooteco.subway.line.LineAcceptanceTest.지하철_노선_등록되어_있음;
-import static wooteco.subway.line.LineAcceptanceTest.지하철_노선_조회_요청;
 import static wooteco.subway.path.PathAcceptanceTest.거리_경로_조회_요청;
 import static wooteco.subway.path.PathAcceptanceTest.총_거리가_응답됨;
 import static wooteco.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
@@ -74,7 +72,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> response = 지하철_구간_생성_요청(신분당선, 강남역, 광교역, 3);
 
         // then
-        지하철_구간_등록_실패됨(response);
+        assertBadRequest(response);
     }
 
     @DisplayName("지하철 노선에 등록되지 않은 역을 기준으로 등록한다.")
@@ -84,7 +82,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> response = 지하철_구간_생성_요청(신분당선, 정자역, 양재역, 3);
 
         // then
-        지하철_구간_등록_실패됨(response);
+        assertBadRequest(response);
     }
 
     @DisplayName("지하철 노선에 등록된 지하철역을 제외한다.")
@@ -108,10 +106,10 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> removeResponse = 지하철_노선에_지하철역_제외_요청(신분당선, 강남역);
 
         // then
-        지하철_노선에_지하철역_제외_실패됨(removeResponse);
+        assertBadRequest(removeResponse);
     }
 
-    @DisplayName("지하철 구간 ")
+    @DisplayName("지하철 구간 거리를 수정한다")
     @Test
     void updateDistance() {
         SectionDistanceRequest sectionDistanceRequest = new SectionDistanceRequest(20);
@@ -122,16 +120,16 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         총_거리가_응답됨(pathResponse, sectionDistanceRequest.getDistance());
     }
 
-    private static ExtractableResponse<Response> 지하철_노선에_구간_거리_수정_요청(
+    private ExtractableResponse<Response> 지하철_노선에_구간_거리_수정_요청(
             final Long id,
-            final StationResponse 강남역,
-            final StationResponse 광교역,
+            final StationResponse upStationResponse,
+            final StationResponse downStationResponse,
             final SectionDistanceRequest sectionDistanceRequest) {
 
         return RestAssured
                 .given().log().all()
-                .queryParam("upStationId", 강남역.getId())
-                .queryParam("downStationId", 광교역.getId())
+                .queryParam("upStationId", upStationResponse.getId())
+                .queryParam("downStationId", downStationResponse.getId())
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(sectionDistanceRequest)
                 .when().put("/lines/{lineId}/sections", id)
@@ -139,23 +137,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
                 .extract();
     }
 
-    public static void 지하철_구간_등록되어_있음(LineResponse lineResponse, StationResponse upStation, StationResponse downStation, int distance) {
-        지하철_구간_생성_요청(lineResponse, upStation, downStation, distance);
-    }
-
-    public static ExtractableResponse<Response> 지하철_구간_생성_요청(LineResponse line, StationResponse upStation, StationResponse downStation, int distance) {
-        SectionRequest sectionRequest = new SectionRequest(upStation.getId(), downStation.getId(), distance);
-
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(sectionRequest)
-                .when().post("/lines/{lineId}/sections", line.getId())
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 지하철_노선에_지하철역_순서_정렬됨(ExtractableResponse<Response> response, List<StationResponse> expectedStations) {
+    private void 지하철_노선에_지하철역_순서_정렬됨(ExtractableResponse<Response> response, List<StationResponse> expectedStations) {
         LineResponse line = response.as(LineResponse.class);
         List<Long> stationIds = line.getStations().stream()
                 .map(it -> it.getId())
@@ -168,7 +150,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         assertThat(stationIds).containsExactlyElementsOf(expectedStationIds);
     }
 
-    public static ExtractableResponse<Response> 지하철_노선에_지하철역_제외_요청(LineResponse line, StationResponse station) {
+    private ExtractableResponse<Response> 지하철_노선에_지하철역_제외_요청(LineResponse line, StationResponse station) {
         return RestAssured
                 .given().log().all()
                 .when().delete("/lines/{lineId}/sections?stationId={stationId}", line.getId(), station.getId())
@@ -176,7 +158,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
                 .extract();
     }
 
-    public static void 지하철_구간_생성됨(ExtractableResponse<Response> response) {
+    private void 지하철_구간_생성됨(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 
@@ -186,17 +168,13 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         지하철_노선에_지하철역_순서_정렬됨(response, stationResponses);
     }
 
-    public static void 지하철_노선에_지하철역_제외됨(ExtractableResponse<Response> result, LineResponse lineResponse, List<StationResponse> stationResponses) {
+    private void 지하철_노선에_지하철역_제외됨(ExtractableResponse<Response> result, LineResponse lineResponse, List<StationResponse> stationResponses) {
         assertThat(result.statusCode()).isEqualTo(HttpStatus.OK.value());
         ExtractableResponse<Response> response = 지하철_노선_조회_요청(lineResponse);
         지하철_노선에_지하철역_순서_정렬됨(response, stationResponses);
     }
 
-    public static void 지하철_구간_등록_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    public static void 지하철_노선에_지하철역_제외_실패됨(ExtractableResponse<Response> response) {
+    private void assertBadRequest(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 }
