+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -75,12 +75,6 @@ public class StationAcceptanceTest extends AcceptanceTest {
         지하철역_목록_노선_포함_조회(response, Arrays.asList(stationLineResponse1, stationLineResponse2));
     }
 
-    private void 지하철역_목록_노선_포함_조회(ExtractableResponse<Response> response, List<StationLineResponse> expectedResponse) {
-        assertThat(Arrays.asList(response.as(StationLineResponse[].class)))
-                .usingRecursiveComparison()
-                .isEqualTo(expectedResponse);
-    }
-
     @DisplayName("지하철역을 제거한다.")
     @Test
     void deleteStation() {
@@ -94,23 +88,7 @@ public class StationAcceptanceTest extends AcceptanceTest {
         지하철역_삭제됨(response);
     }
 
-    public static StationResponse 지하철역_등록되어_있음(String name) {
-        return 지하철역_생성_요청(name).as(StationResponse.class);
-    }
-
-    public static ExtractableResponse<Response> 지하철역_생성_요청(String name) {
-        StationRequest stationRequest = new StationRequest(name);
-
-        return RestAssured
-                .given().log().all()
-                .body(stationRequest)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/stations")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철역_목록_조회_요청() {
+    private ExtractableResponse<Response> 지하철역_목록_조회_요청() {
         return RestAssured
                 .given().log().all()
                 .when().get("/stations")
@@ -118,7 +96,13 @@ public class StationAcceptanceTest extends AcceptanceTest {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 지하철역_제거_요청(StationResponse stationResponse) {
+    private void 지하철역_목록_노선_포함_조회(ExtractableResponse<Response> response, List<StationLineResponse> expectedResponse) {
+        assertThat(Arrays.asList(response.as(StationLineResponse[].class)))
+                .usingRecursiveComparison()
+                .isEqualTo(expectedResponse);
+    }
+
+    private ExtractableResponse<Response> 지하철역_제거_요청(StationResponse stationResponse) {
         return RestAssured
                 .given().log().all()
                 .when().delete("/stations/" + stationResponse.getId())
@@ -126,32 +110,20 @@ public class StationAcceptanceTest extends AcceptanceTest {
                 .extract();
     }
 
-    public static void 지하철역_생성됨(ExtractableResponse response) {
+    private void 지하철역_생성됨(ExtractableResponse response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
         assertThat(response.header("Location")).isNotBlank();
     }
 
-    public static void 지하철역_생성_실패됨(ExtractableResponse<Response> response) {
+    private void 지하철역_생성_실패됨(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
-    public static void 지하철역_목록_응답됨(ExtractableResponse<Response> response) {
+    private void 지하철역_목록_응답됨(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 
-    public static void 지하철역_삭제됨(ExtractableResponse<Response> response) {
+    private void 지하철역_삭제됨(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
-    public static void 지하철역_목록_포함됨(ExtractableResponse<Response> response, List<StationResponse> createdResponses) {
-        List<Long> expectedLineIds = createdResponses.stream()
-                .map(it -> it.getId())
-                .collect(Collectors.toList());
-
-        List<Long> resultLineIds = response.jsonPath().getList(".", StationResponse.class).stream()
-                .map(StationResponse::getId)
-                .collect(Collectors.toList());
-
-        assertThat(resultLineIds).containsAll(expectedLineIds);
     }
 }
